searchSystem/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-import-export==3.2.0
django-simpleui==2023.3.1
et-xmlfile==1.1.0
fonttools==4.39.3
idna==3.4
jieba==0.42.1
kiwisolver==1.4.4
MarkupPy==1.14
matplotlib==3.7.1
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
tzdata==2023.3
urllib3==2.0.2
wordcloud==1.9.1.1
xlrd==2.0.1
xlwt==1.3.0
